Separate context provider, context creator and useContext wrapper hook
