+++ src/contexts/HomeContextDefinition.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+import type { HomeContextType } from '../types/home'
+
+export const HomeContext = createContext<HomeContextType | undefined>(undefined)
